#393 Фильтрация Locality через делегат
